Update StronglyTyped Id NSwag OpenApi output adding the Format
