+++ application/shared-kernel/ApiCore/ApiCoreConfiguration.cs
@@ -7,14 +7,13 @@ using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
-using NJsonSchema;
 using NJsonSchema.Generation;
 using PlatformPlatform.SharedKernel.ApiCore.Aspire;
 using PlatformPlatform.SharedKernel.ApiCore.Endpoints;
 using PlatformPlatform.SharedKernel.ApiCore.Filters;
 using PlatformPlatform.SharedKernel.ApiCore.Middleware;
+using PlatformPlatform.SharedKernel.ApiCore.SchemaProcessor;
 using PlatformPlatform.SharedKernel.ApiCore.SinglePageApp;
-using PlatformPlatform.SharedKernel.DomainCore.Identity;
 
 namespace PlatformPlatform.SharedKernel.ApiCore;
 
@@ -70,22 +69,7 @@ public static class ApiCoreConfiguration
                 options.SerializerOptions.Converters.Add(new JsonStringEnumConverter());
                 options.SerializerOptions.PropertyNamingPolicy = JsonNamingPolicy.CamelCase;
 
-                settings.PostProcess = document =>
-                {
-                    // Find all strongly typed IDs
-                    var stronglyTypedIdNames = domainAssembly.GetTypes()
-                        .Where(t => typeof(IStronglyTypedId).IsAssignableFrom(t))
-                        .Select(t => t.Name)
-                        .ToList();
-
-                    // Ensure the Swagger UI to correctly display strongly typed IDs as plain text instead of complex objects
-                    foreach (var stronglyTypedIdName in stronglyTypedIdNames)
-                    {
-                        var schema = document.Definitions[stronglyTypedIdName];
-                        schema.Type = JsonObjectType.String;
-                        schema.Properties.Clear();
-                    }
-                };
+                settings.DocumentProcessors.Add(new StronglyTypedDocumentProcessor(domainAssembly));
             }
         );
